requirements.txt
# Core Django dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Environment and configuration
python-decouple==3.8

# Database
psycopg==3.1.12

# Caching and sessions
redis==5.0.1
django-redis==5.4.0

# Security and authentication
PyJWT==2.8.0
cryptography==41.0.7
argon2-cffi==23.1.0

# Background tasks
celery==5.3.4

# Monitoring and logging
structlog==23.2.0
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Development dependencies
django-extensions==3.2.3
django-debug-toolbar==4.2.0
whitenoise==6.6.0